Move:  都道府県のモデルを別ファイルに分けた
