+++ PrefectureChemistry/ResultView.swift
@@ -7,20 +7,6 @@
 
 import SwiftUI
 
-class Prefecture: Codable {
-    let name: String
-    let capital: String
-    let citizen_day: MonthDay?
-    let has_coast_line: Bool
-    let logo_url: String
-    let brief: String
-}
-
-class MonthDay: Codable {
-    let month: Int
-    let day: Int
-}
-
 struct ResultView: View {
     let prefecture: Prefecture
     @State var image: UIImage?
